다 갈아엎고 게시판 CRUD 세팅
글 작성, 조회, 수정, 삭제 가능.
이제 회원가입 기능을 구현해서 자신이 작성한 글만 보이게 하는 게 필요함.

+++ src/main/java/hongik/diary/repository/DiaryRepository.java
@@ -3,5 +3,6 @@ package hongik.diary.repository;
 import hongik.diary.entity.Diary;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface DiaryRepository extends JpaRepository<Diary, Integer> {
+public interface DiaryRepository extends JpaRepository<Diary, Long> {
+
 }
